add files for 0x1E-search_algorithms project

+++ 0x1E-search_algorithms/0-linear.c
@@ -11,14 +11,14 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int i;
+	size_t i;
 
 	if (array == NULL)
 		return (-1);
 
 	for (i = 0; i < size; i++)
 	{
-		printf("value checked array[%d] = [%d]\n", i, array[i]);
+		printf("value checked array[%ld] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 		{
 			return (i);
